Jaaga/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
qrcode==7.4.2
Pillow==10.0.1
requests==2.31.0
bcrypt==4.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
gunicorn==21.2.0
opencv-python==4.8.1.78
pytesseract==0.3.10 
